requirements.txt
streamlit
google-generativeai
python-docx
Pillow
pypdf
opencv-python-headless
opencv